A huge overhaul of the plotting class
For now it has a different class name than the standard plotting class,
at least for now. The new plotting class makes adding new plot far more
trivial and automated, only the executable needs to be added (and it's
most likely one extra line of code).

There are new structs with the plot types (that can be improved
further), each variable has it's ACNET struct, and there is a new Cut
struct too (that wcan be improved further to facilitate extra cuts).

This still requires more work to neat-up the code and make it more
flexible, but it is already far better than the previous plotting class
in classes/plotting.cpp.